Initial commit
The project now has good package scheme. I also use ant to build
and run the project.

TODO: use ant to build jar and zip archive.

UPDATE: the old connect6 history is converted to git's style.
After some hack, the initial commit points to the tip of the old
history as its parent.

NOTE: You can notice that the encoding of the source files is
converted from CP936 to UTF-8.